refact: organize code
